Share the same object
